requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gitdb==4.0.10
GitPython==3.1.32
PyJWT==2.8.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
smmap==5.0.0
sqlparse==0.4.4
